refactor cli server commands
