+++ crates/rulemorph_cli/tests/cli.rs
@@ -25,3 +25,5 @@ include!("cli/limits.rs");
 include!("cli/generate.rs");
 
 include!("cli/api_keys.rs");
+
+include!("cli/server_commands.rs");
